Create letsGuessYourAge.js
This is an easy JS program to know your age using the current year and your current age. I know this program can be improved this one was thought for beginners as a practice code that can be executed in the navigator only. If you want, you can add an HTML structure and style it with CSS.